Changed return value of function which calculate wasted space (disk space occupied by duplicated files)

+++ src/duplicatesfinder.h
@@ -30,7 +30,7 @@ private:
     SizeMap duplicatesMap;
     int _numberOfFiles;
     int _numberOfDuplicates;
-    int _wastedSpace;
+    qint64 _wastedSpace;
     int _strength;
     QString hash(QString fileName,qint64 chunkSize);
 };
